requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==3.0.1
google-generativeai==0.3.2
reportlab==4.0.7
Pillow>=9.0.0
python-escpos>=3.1
pyusb>=1.3.1


